move sample functions ABIPack, CompactPack into memo pkg self; remove sample package reference from memo to minimize dependency
